test(satellite): manually shard test upgrades
